fix: Code refactoring
