+++ frontend/packages/ui/src/stories/Button.stories.tsx
@@ -1,7 +1,6 @@
 import { Meta } from '@storybook/react'
-import { styled } from '@storybook/theming'
 import React from 'react'
-
+import styled from 'styled-components'
 import { BoxLayout } from '..'
 import { Button } from '../Button'
